Add readme, tune arg boundaries to specs, testing
With this commit, a README file is added that answers the prompts in the
project PDF and provides information on all testing files not included
with the project.  All parameters are explained, and valid inputs
elucidated.  Testing files have been added that aim to help the grader
in seeing that the algorithm works as expected.